Fix installation to include media files.
This patch copies the logo images to /usr/share/samba-gtk and if at runtime if it fails to
load the images from the locally directory loads it from there.

Also added icons to the .desktop files and changed setup.py to use absolute paths.

Unfortunately taskbars don't seem to necessarily pick up on the application icons at this stage.